R through Jupyter Notebook
First, I install R inside Juypter Notebook but the problem with this is that the r packages here doesn't seem to be update unlike Rstudio.
Tidygraph should have worked if I were to update dplyr into 0.8.5 but r in anaconda only seems to be support 0.8.0.1.
Next thing I tried is making rmd file first and then convert into ipynb file using rmd2jupyter but the problem with this is that the language crashes even though I upload through kaggle. I'm uploading the code I have used as a reference of guide to SNA and recommend opening through Rstudio.